+++ crates/bindings/Cargo.toml
@@ -6,10 +6,10 @@ license.workspace = true
 authors.workspace = true
 
 [dependencies]
-controller = { path = "../controller" }
-io = { path = "../io" }
-settings = { path = "../settings" }
 config = { version = "0.15.3", features = ["toml"] }
+controller = { path = "../controller" }
 directories = "6.0.0"
+io = { path = "../io" }
 serde = { version = "1.0.219", features = ["derive"] }
+settings = { path = "../settings" }
 toml = "0.9.2"
