+++ crates/controller/Cargo.toml
@@ -6,5 +6,5 @@ license.workspace = true
 authors.workspace = true
 
 [dependencies]
-settings = { path = "../settings" }
 serde = { version = "1.0.219", features = ["derive"] }
+settings = { path = "../settings" }
