+++ crates/joymouse/src/main.rs
@@ -1,3 +1,10 @@
 fn main() {
-  platform::Controller::run()
+  #[cfg(not(windows))]
+  {
+    platform::linux::Controller::run()
+  }
+  #[cfg(windows)]
+  {
+    println!("Windows is not supported yet. Stay tuned :)");
+  }
 }
